% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustrPairs.R
\name{ClustrPairs}
\alias{ClustrPairs}
\title{Connected Components from connected pairs}
\description{
This script takes in "Pair1  Pair2 Similarity_score" (tab seperated) and cluster the pairs which have threshold similirity cutoff. The final output contains unique list of pairs and their corresponding cluster number
}
\examples{
ClustrPairs("connected_pairs.txt")
}
\keyword{connected_pairs}
